Tried and failed to get the text to update
I could not figure out how to get the text to update instead of just displaying "##".
Ran into issues where it said it could not convert int to string; once I got rid of that error, the text still is not updating.
Will try again later and hopefully get it done and uploaded to Itch before the deadline.